refactor(profile): move settings components to features
- Move settings from app to features/profile/settings/ui
- Update imports in page.tsx
- App should only compose, not contain components

Related: SCRUM-119
