+++ src/entities/profile/model/index.ts
@@ -1,0 +1,2 @@
+export * from './domain/types'
+export * from './type-guards'
