+++ src/features/profile/settings/index.ts
@@ -1,0 +1,4 @@
+export { AccountManagement } from './ui/AccountManagement'
+export { Devices } from './ui/Devices'
+export { GeneralInfo } from './ui/General/GeneralInfo'
+export { Payments } from './ui/Payments'
